Extending command line args to include flow statements and the chalk library.
